Resolve a plan file from Plan itself
The resolver belongs with the class that already reads plan files, not
in a file of its own: `Plan.resolvePath()` beside `Plan.fromMarkdown()`,
with its tests in the Plan suite.
